requirements.txt
imutils==0.5.3
numpy==1.19.1
opencv-python==3.4.11.41
pyserial==3.4